+++ ui/src/app/core/accessories/types/matter/door-lock/door-lock.manage.component.ts
@@ -1,12 +1,9 @@
-import { ChangeDetectionStrategy, ChangeDetectorRef, Component, DestroyRef, inject, OnInit } from '@angular/core'
-import { takeUntilDestroyed } from '@angular/core/rxjs-interop'
-import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap/modal'
+import { ChangeDetectionStrategy, Component, inject } from '@angular/core'
 import { TranslatePipe } from '@ngx-translate/core'
 import { ToastrService } from 'ngx-toastr'
 
-import { ServiceTypeX } from '@/app/core/accessories/accessories.interfaces'
-import { AccessoriesService } from '@/app/core/accessories/accessories.service'
-import { ACCESSORY_MANAGE_MODAL_DATA } from '@/app/core/accessories/types/base-manage.component'
+import { BaseManageComponent } from '@/app/core/accessories/types/base-manage.component'
+import { DoorLockState } from '@/app/core/accessories/types/matter/matter-device.constants'
 import { getDoorLockState, setDoorLockState } from '@/app/core/accessories/types/matter/matter-device.utils'
 
 @Component({
@@ -18,60 +15,16 @@ import { getDoorLockState, setDoorLockState } from '@/app/core/accessories/types
   templateUrl: './door-lock.manage.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class DoorLockManageComponent implements OnInit {
-  protected destroyRef = inject(DestroyRef)
-  protected $activeModal = inject(NgbActiveModal)
-  protected cdr = inject(ChangeDetectorRef)
+export class DoorLockManageComponent extends BaseManageComponent {
   private $toastr = inject(ToastrService)
-
-  // Inject modal data using modern DI pattern
-  private modalData = inject(ACCESSORY_MANAGE_MODAL_DATA)
-
-  // Public properties for component use (accessed by templates)
-  public service!: ServiceTypeX
-  public $accessories!: AccessoriesService
 
   public targetMode: number
 
-  public ngOnInit() {
-    // Null safety check
-    if (!this.modalData.service || !this.modalData.$accessories) {
-      console.error('DoorLockManageComponent: service or $accessories not provided')
-      this.$activeModal.dismiss('Missing required data')
-      return
-    }
-
-    // Store in public properties (same object references)
-    this.service = this.modalData.service
-    this.$accessories = this.modalData.$accessories
-
-    this.setupComponent()
-    this.subscribeToAccessoryUpdates()
-  }
-
-  public dismissModal() {
-    this.$activeModal.dismiss('Dismiss')
-  }
-
-  private setupComponent() {
+  protected setupComponent() {
     this.targetMode = getDoorLockState(this.service)
   }
 
-  private subscribeToAccessoryUpdates() {
-    if (this.$accessories) {
-      this.$accessories.accessoryData.pipe(takeUntilDestroyed(this.destroyRef)).subscribe(() => {
-        // Update service reference to get latest data (zoneless Angular compatibility)
-        const updatedService = this.$accessories.accessories.services.find(s => s.uniqueId === this.service.uniqueId)
-        if (updatedService) {
-          this.service = updatedService
-        }
-        this.handleAccessoryUpdate()
-        this.cdr.markForCheck()
-      })
-    }
-  }
-
-  private handleAccessoryUpdate() {
+  protected handleAccessoryUpdate() {
     this.targetMode = getDoorLockState(this.service)
   }
 
@@ -82,20 +35,15 @@ export class DoorLockManageComponent implements OnInit {
       this.targetMode = value
       this.cdr.markForCheck()
 
-      const locked = value === 1
+      const locked = value === DoorLockState.Locked
       await setDoorLockState(this.service, locked)
 
       this.blurTarget(event)
     } catch (error) {
-      this.$toastr.error(`Failed to ${value === 1 ? 'lock' : 'unlock'} door`, 'Error')
+      this.$toastr.error(`Failed to ${value === DoorLockState.Locked ? 'lock' : 'unlock'} door`, 'Error')
       // Revert to previous state on error
       this.targetMode = previousMode
       this.cdr.markForCheck()
     }
-  }
-
-  protected blurTarget(event: MouseEvent) {
-    const target = event.target as HTMLButtonElement
-    target.blur()
   }
 }