+++ ui/src/app/core/accessories/types/matter/window-covering/window-covering.manage.component.ts
@@ -1,16 +1,11 @@
-import { ChangeDetectionStrategy, ChangeDetectorRef, Component, DestroyRef, inject, OnInit } from '@angular/core'
-import { takeUntilDestroyed } from '@angular/core/rxjs-interop'
+import { ChangeDetectionStrategy, Component, inject } from '@angular/core'
 import { FormsModule } from '@angular/forms'
-import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap/modal'
 import { TranslatePipe } from '@ngx-translate/core'
 import { NouisliderComponent } from 'ng2-nouislider'
 import { ToastrService } from 'ngx-toastr'
 import { Subject } from 'rxjs'
-import { debounceTime } from 'rxjs/operators'
 
-import { ServiceTypeX } from '@/app/core/accessories/accessories.interfaces'
-import { AccessoriesService } from '@/app/core/accessories/accessories.service'
-import { ACCESSORY_MANAGE_MODAL_DATA } from '@/app/core/accessories/types/base-manage.component'
+import { BaseManageComponent } from '@/app/core/accessories/types/base-manage.component'
 import { getWindowCoveringPercentage, setWindowCoveringPosition } from '@/app/core/accessories/types/matter/matter-device.utils'
 
 @Component({
@@ -24,18 +19,8 @@ import { getWindowCoveringPercentage, setWindowCoveringPosition } from '@/app/co
   templateUrl: './window-covering.manage.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class WindowCoveringManageComponent implements OnInit {
-  protected destroyRef = inject(DestroyRef)
-  protected $activeModal = inject(NgbActiveModal)
-  protected cdr = inject(ChangeDetectorRef)
+export class WindowCoveringManageComponent extends BaseManageComponent {
   private $toastr = inject(ToastrService)
-
-  // Inject modal data using modern DI pattern
-  private modalData = inject(ACCESSORY_MANAGE_MODAL_DATA)
-
-  // Public properties for component use (accessed by templates)
-  public service!: ServiceTypeX
-  public $accessories!: AccessoriesService
 
   public targetPositionChanged: Subject<number> = new Subject<number>()
   public targetPosition: {
@@ -45,27 +30,7 @@ export class WindowCoveringManageComponent implements OnInit {
     step: number
   }
 
-  public ngOnInit() {
-    // Null safety check
-    if (!this.modalData.service || !this.modalData.$accessories) {
-      console.error('WindowCoveringManageComponent: service or $accessories not provided')
-      this.$activeModal.dismiss('Missing required data')
-      return
-    }
-
-    // Store in public properties (same object references)
-    this.service = this.modalData.service
-    this.$accessories = this.modalData.$accessories
-
-    this.setupComponent()
-    this.subscribeToAccessoryUpdates()
-  }
-
-  public dismissModal() {
-    this.$activeModal.dismiss('Dismiss')
-  }
-
-  private setupComponent() {
+  protected setupComponent() {
     this.loadTargetPosition()
 
     // Subscribe to target position changes with debounce
@@ -85,21 +50,7 @@ export class WindowCoveringManageComponent implements OnInit {
     )
   }
 
-  private subscribeToAccessoryUpdates() {
-    if (this.$accessories) {
-      this.$accessories.accessoryData.pipe(takeUntilDestroyed(this.destroyRef)).subscribe(() => {
-        // Update service reference to get latest data (zoneless Angular compatibility)
-        const updatedService = this.$accessories.accessories.services.find(s => s.uniqueId === this.service.uniqueId)
-        if (updatedService) {
-          this.service = updatedService
-        }
-        this.handleAccessoryUpdate()
-        this.cdr.markForCheck()
-      })
-    }
-  }
-
-  private handleAccessoryUpdate() {
+  protected handleAccessoryUpdate() {
     if (this.targetPosition) {
       this.targetPosition.value = getWindowCoveringPercentage(this.service)
     }
@@ -107,25 +58,6 @@ export class WindowCoveringManageComponent implements OnInit {
 
   public onTargetPositionChange() {
     this.targetPositionChanged.next(this.targetPosition.value)
-  }
-
-  private createDebouncedSubscription<T>(
-    subject$: Subject<T>,
-    callback: (value: T) => void,
-    debounceMs: number = 500,
-  ) {
-    subject$
-      .pipe(debounceTime(debounceMs), takeUntilDestroyed(this.destroyRef))
-      .subscribe(callback)
-  }
-
-  private applySliderGradient(gradient: string, selector: string = '.noUi-target') {
-    requestAnimationFrame(() => {
-      const sliderElements = document.querySelectorAll<HTMLElement>(selector)
-      sliderElements.forEach((sliderElement) => {
-        sliderElement.style.background = gradient
-      })
-    })
   }
 
   private loadTargetPosition() {