+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - info widget - remove margin on small table (#2727) (@dnicolson)
 - feat: settings UI reorganization
 - feat: port overview modal
+- feat: matter device management improvements
 - upgrade `xterm` to `v6` + style improvements
 - added option: global terminal/log window settings
 - use new font-awesome `ko-fi` icon for donate modal